Added tests and improved the logic

+++ app/controllers/locks_controller.rb
@@ -8,15 +8,23 @@ class LocksController < RestController
   before_action :current_user
 
   LOCK_NOT_FOUND = "Lock not found"
+  ID_FIELD_ALLOWED_CHARACTERS = /\A[a-zA-Z0-9+\-_]+\z/
 
   def create
-    logger.info("POST /locks/:account started")
+    logger.debug("POST /locks/:account started")
     authorize(:create)
+
+    unless params[:ttl].is_a? Integer and params[:ttl] > 0
+      raise ApplicationController::BadRequest, "invalid 'ttl' parameter. Must be a positive number"
+    end
+
+    unless params[:id].match?(ID_FIELD_ALLOWED_CHARACTERS)
+      raise ApplicationController::BadRequest, "invalid 'id' parameter. Only the following characters are supported: A-Z, a-z, 0-9, +, -, and _"
+    end
 
     delete_lock_if_expired(params[:account], params[:id])
 
-    current_timestamp = Time.now
-    expires_at = current_timestamp + params[:ttl]
+    expires_at = Lock.db.get(Sequel.lit("CURRENT_TIMESTAMP + interval ?", "#{params[:ttl]} second"))
     lock = Lock.new(lock_id: params[:id], account: params[:account],
                     owner: params[:owner],
                     modified_at: Sequel::CURRENT_TIMESTAMP,
@@ -24,18 +32,20 @@ class LocksController < RestController
     lock.save
 
     render(json: lock.as_json, status: :created)
-    logger.info("POST /locks/:account ended successfully")
+    logger.debug("POST /locks/:account ended successfully")
   rescue Sequel::UniqueConstraintViolation => e
     logger.error("The lock [#{params[:id]}] already exists")
+    logger.debug("POST /locks/:account ended successfully")
     raise Exceptions::RecordExists.new("lock", params[:id])
   end
 
   def get
-    logger.info("GET /locks/:account/*identifier started")
+    logger.debug("GET /locks/:account/*identifier started")
     authorize(:read)
 
+    delete_lock_if_expired(params[:account], params[:identifier])
     lock = get_lock_from_db(params[:account], params[:identifier])
-    if lock and delete_lock_if_expired(params[:account], params[:identifier]) == 0
+    if lock
       render(json: lock.as_json, status: :ok)
     else
       render(json: {
@@ -52,18 +62,18 @@ class LocksController < RestController
       }, status: :not_found)
     end
 
-    logger.info("GET /locks/:account/*identifier ended successfully")
+    logger.debug("GET /locks/:account/*identifier ended successfully")
   end
 
   def update
-    logger.info("PATCH /locks/:account/*identifier started")
+    logger.debug("PATCH /locks/:account/*identifier started")
     authorize(:update)
 
+    delete_lock_if_expired(params[:account], params[:identifier])
     lock = get_lock_from_db(params[:account], params[:identifier])
-    if lock and delete_lock_if_expired(params[:account], params[:identifier]) == 0
-      current_timestamp = Time.now
-      expires_at = current_timestamp + params[:ttl]
-      Lock.where(account: account, lock_id: params[:identifier]).update(modified_at: current_timestamp, expires_at: expires_at)
+    if lock
+      expires_at = Lock.db.get(Sequel.lit("CURRENT_TIMESTAMP + interval ?", "#{params[:ttl]} second"))
+      Lock.where(account: account, lock_id: params[:identifier]).update(modified_at: Sequel::CURRENT_TIMESTAMP, expires_at: expires_at)
       render(json: lock.as_json, status: :ok)
     else
       render(json: {
@@ -79,15 +89,16 @@ class LocksController < RestController
         }
       }, status: :not_found)
     end
-    logger.info("PATCH /locks/:account/*identifier ended successfully")
+    logger.debug("PATCH /locks/:account/*identifier ended successfully")
   end
 
   def delete
-    logger.info("DELETE /locks/:account/*identifier started")
+    logger.debug("DELETE /locks/:account/*identifier started")
     authorize(:delete)
 
+    delete_lock_if_expired(params[:account], params[:identifier])
     lock = get_lock_from_db(params[:account], params[:identifier])
-    if lock and delete_lock_if_expired(params[:account], params[:identifier]) == 0
+    if lock
       lock.delete
       head :ok
     else
@@ -104,7 +115,7 @@ class LocksController < RestController
         }
       }, status: :not_found)
     end
-    logger.info("DELETE /locks/:account/*identifier ended successfully")
+    logger.debug("DELETE /locks/:account/*identifier ended successfully")
   end
 
   private
